refactor(users): import usecases from index
